Refactors and styles Settings component
