+++ src/Settings/Settings.css
@@ -2,16 +2,26 @@
   color: white;
   font-size: 24px;
   margin-bottom: 20px;
+  margin-left: 20px;
 }
 
 .Settings h2 {
   color: white;
   font-size: 16px;
-  margin-bottom: 5px;
+  margin-bottom: 10px;
 }
 
+.Settings > div {
+  background-color: rgb(0, 0, 0, 0.6);
+  margin: 0 10px 20px 10px;
+  padding: 20px;
+  max-width: fit-content;
+  border-radius: 5px;
+}
+
+#background,
 #config-url {
-  width: 33%;
+  width: 400px;
 }
 
 .Settings {
@@ -19,7 +29,8 @@
   color: white;
 }
 
-.refresh-icon {
+.refresh-icon,
+.download-icon {
   color: white;
   margin-left: 5px;
 }
